Protecting Video Routes

+++ controllers/videoController.js
@@ -59,7 +59,6 @@ export const videoDetail = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id).populate("creator");
-        console.log(video);
 
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
@@ -74,9 +73,12 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
-        console.log(error);
         res.redirect(routes.home);
     }
 };
@@ -102,7 +104,12 @@ export const deleteVideo = async (req, res) => {
         params: { id }
     } = req;
     try {
-        await Video.findByIdAndRemove({ _id: id });
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findByIdAndRemove({ _id: id });
+        }
     } catch (error) {
         console.log(error);
     }
